+++ src/Interp/Directives.hs
@@ -1,14 +1,16 @@
 {-# LANGUAGE NoImplicitPrelude #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module Interp.Directives where
 
 import           RIO
+import qualified Data.Text.Lazy                as TL
+import           Utils.IO                       ( putStr )
 import           Utils.Types.App                ( Interp )
 import           Utils.Types                    ( Scope(scope)
                                                 , globalL
                                                 )
-import           Utils.Interp                   ( printString
-                                                , putStore
+import           Utils.Interp                   ( putStore
                                                 , withScopes
                                                 )
 import           Interp.Directives.Parser       ( parseDrct
@@ -18,19 +20,19 @@ import           Data.Map                       ( delete
                                                 , empty
                                                 )
 
-isDirective :: String -> Bool
-isDirective s = ':' `elem` s
+isDirective :: TL.Text -> Bool
+isDirective s = ":" `TL.isPrefixOf` s
 
-execute :: String -> Interp a ()
+execute :: TL.Text -> Interp a ()
 execute str = case parseDrct str of
-  Left  err  -> printString (err ++ "\n")
+  Left  err  -> putStr $ TL.pack err <> "\n"
   Right drct -> exec drct
 
 exec :: Drct -> Interp a ()
 exec Exit  = putStore (Left ())
 exec Clear = withScopes $ (over . scope) globalL (const empty)
 exec Help =
-  printString
+  putStr
     "PiG interpreter directives: \n\
     \:help | :h - display this message\n\
     \:exit | :e or Ctrl+d - leave the interpreter\n\