+++ src/Interp/Directives/Parser.hs
@@ -1,14 +1,15 @@
 {-# LANGUAGE NoImplicitPrelude #-}
+{-# LANGUAGE OverloadedStrings #-}
 module Interp.Directives.Parser where
 
 import           RIO                     hiding ( many
                                                 , optional
                                                 , try
-                                                , (<|>)
                                                 )
+import qualified Data.Text.Lazy                as TL
 import           Utils.Types                    ( Var )
-import           Text.Parsec
-import           Text.Parsec.String             ( Parser )
+
+import           Data.Attoparsec.Text.Lazy
 
 data Drct
   = Exit
@@ -16,16 +17,16 @@ data Drct
   | Help
   | Rm Var
 
-parseDrct :: String -> Either String Drct
-parseDrct s = case parse drctParser "PiG REPL" s of
-  Left  err  -> Left $ show err
-  Right drct -> Right drct
+parseDrct :: TL.Text -> Either String Drct
+parseDrct s = case parse drctParser s of
+  Fail _ _ err -> Left err
+  Done _ drct  -> Right drct
 
 drctParser :: Parser Drct
 drctParser =
-  many (char ' ')
+  many' (char ' ')
     *>  try (string ":e" $> Exit)
     <|> try (string ":h" $> Help)
     <|> try (string ":c" $> Clear)
-    <|> try (string ":rm " *> (Rm <$> many anyToken))
+    <|> try (string ":rm " *> (Rm <$> takeLazyText))
     <?> "PiG directives"